make comic[last_entries] only update if new entries
comic['last_entries'] is meant to be the 3 most recent entries, to
ensure that if the most recent one is depublished, we will still see the
2nd most recent and not update.

However, previously worker.py would always add the most recent entry to
last_entry, even if it had already been added. The new behaviour is to
add only *new* entries, from least to most recent.