Turn off SSL checks
There was an issue for a week where the script would fail to connect to
Webtoons, due to SSL checks. However, the site itself still worked, so
the certificate must have been fine, and the issue must have been on
this end.

This "fixes" that by just turning off all SSL checks. It seems to work fine as
far as I can tell, but may have issues that did not appear in testing.

I live in fear