Use caching headers when possible
This sets the default caching policy to no-cache, hopefully forcing any
intermediary caching servers to check for freshness.

It also uses ETag and Last-Modified headers when those exist (sadly, not
often). This necessitates adding a few new functions and modifying the
database schema a little, but was surprisingly simple.

I've realised that's tempting fate, so fingers crossed.