Change db connection from global to parameter
This is necessary so that inner functions still have access to the db
when the main function is called from another file. This is obviously
an ugly solution. I hope to find a better one eventually, and also
ensure that all db changes occur at once to prevent inconsistent states.