Fix mysterious ClientPayloadError
This error only appeared on the cloned version of the repo for reasons
I don't understand. As far as I can tell the root issue is a bug in
aiohttp [here](https://github.com/aio-libs/aiohttp/issues/4581) that
no-one's been able to isolate for 3 years. Let's hope this workaround
works.