Merge all MongoDB collections
This change fundamentally reworks a lot of the logic in pursuit of
making it slightly easier to add a comic to the database. It needs a lot of
and is very ambitious and risky. This was the smallest atomic commit I
could make since every intermediate state was full of errors.