Make mypy ignore the scripts directory
It's already in .gitignore so this won't make any sense to people
reading this, but the scripts directory contains a bunch of messy
scripts that do things that are definitely not typed correctly but that
it's not worth it to fix, often because they migrate to different
database schemas, so they have a different type for `Comic` items than
the current one.

This change lets me ignore those files when i run the mypy command, which
should make the mypy output a lot more usable.