Post all of a comic's new pages in one message
This change changes the structure of the code a lot, and also breaks a
lot of the tests. The ones that can easily be updated have been, and the
ones that can't have been disabled for now. I'm trying to push this
through to staging quickly so I can get a good idea of what this looks
like in practice.