Add a failing test for a strange edge case
Currently if an RSS feed without, e.g., <pubDate>s adds them to all
entries, then the checker will not think there are any new entries.
This still works even when a new entry is added to the feed.
However, the entry after that one will cause every old update in the
feed to be posted again.

Adding a failing test case in advance of fixing the bug.