Avoid use of `+` on `Sequence`s
This is a type error under mypy, because `Sequence` doesn't implement
`__add__`.
Both operands are actually `list`s, but `list`s are mutable and so
invariant, so annotating them as such would cause type errors.
We don't actually need mutability, but there's no ABC for immutable
sequences that can be concatenated, so let's just give up on trying to
use `+` and spread the lists instead.