Replace deepcopy on `TypedDicts` with constructor
It's possible to use the constructor for a typeddict to create a copy
with certain keys over-ridden, just like with normal dicts.
This should be slightly faster, since it avoids creating a full deepcopy,
and doesn't copy things that are just immediately over-ridden.

This only creates a shallow copy, but every copy is immediately inserted
into a database, so the aliasing can't cause any issues.